hw3-coding-fcn/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
backcall=0.1.0=py36_0
blas=1.0=mkl
bleach=3.1.0=py36_0
ca-certificates=2019.1.23=0
certifi=2018.11.29=py36_0
cffi=1.11.5=py36he75722e_1
cuda80=1.0=h205658b_0
cycler=0.10.0=py36_0
dbus=1.13.6=h746ee38_0
decorator=4.3.2=py36_0
entrypoints=0.3=py36_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
glib=2.56.2=hd408876_0
gmp=6.1.2=h6c8ec71_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=h9c2bf20_1
intel-openmp=2019.1=144
ipykernel=5.1.0=py36h39e3cac_0
ipython=7.2.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.4.2=py36_0
jedi=0.13.2=py36_0
jinja2=2.10=py36_0
jpeg=9b=h024ee3a_2
jsonschema=2.6.0=py36h006f8b5_0
jupyter=1.0.0=py36_7
jupyter_client=5.2.4=py36_0
jupyter_console=6.0.0=py36_0
jupyter_core=4.4.0=py36_0
kiwisolver=1.0.1=py36hf484d3e_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.36=hbc83047_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=he19cac6_0
markupsafe=1.1.0=py36h7b6447c_0
matplotlib=3.0.2=py36h5429711_0
mistune=0.8.4=py36h7b6447c_0
mkl=2019.1=144
mkl_fft=1.0.10=py36ha843d7b_0
mkl_random=1.0.2=py36hd81dba3_0
nbconvert=5.3.1=py36_0
nbformat=4.4.0=py36_0
ncurses=6.1=he6710b0_1
ninja=1.8.2=py36h6bb024c_1
notebook=5.7.4=py36_0
numpy=1.15.4=py36h7e9f1db_0
numpy-base=1.15.4=py36hde5b4d6_0
olefile=0.46=py36_0
openssl=1.1.1a=h7b6447c_0
pandas=0.24.1=py36he6710b0_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
parso=0.3.2=py36_0
pcre=8.42=h439df22_0
pexpect=4.6.0=py36_0
pickleshare=0.7.5=py36_0
pillow=5.4.1=py36h34e0f95_0
pip=18.1=py36_0
prometheus_client=0.5.0=py36_0
prompt_toolkit=2.0.8=py_0
ptyprocess=0.6.0=py36_0
pycparser=2.19=py36_0
pygments=2.3.1=py36_0
pyparsing=2.3.1=py36_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.8=h0371630_0
python-dateutil=2.7.5=py36_0
pytorch=1.0.0=py3.6_cuda8.0.61_cudnn7.1.2_1
pytz=2018.9=py36_0
pyzmq=17.1.2=py36he6710b0_2
qt=5.9.7=h5867ecd_1
qtconsole=4.4.3=py36_0
readline=7.0=h7b6447c_5
scikit-learn=0.20.2=py36hd81dba3_0
scipy=1.2.0=py36h7c811a0_0
send2trash=1.5.0=py36_0
setuptools=40.6.3=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=py36_0
sqlite=3.26.0=h7b6447c_0
terminado=0.8.1=py36_1
testpath=0.4.2=py36_0
tk=8.6.8=hbc83047_0
torchvision=0.2.1=py_2
tornado=5.1.1=py36h7b6447c_0
traitlets=4.3.2=py36_0
wcwidth=0.1.7=py36hdf4376a_0
webencodings=0.5.1=py36_1
wheel=0.32.3=py36_0
widgetsnbextension=3.4.2=py36_0
xz=5.2.4=h14c3975_4
zeromq=4.3.1=he6710b0_3
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
